test: gate DefinePlugin optional-chaining cases behind supportsOptionalChaining
The bundle emits `?.` syntax, which Node 10/12 cannot parse, so move the
optional-chaining and optional-call assertions into separate config cases
guarded by a supportsOptionalChaining test.filter. The main define-plugin
case keeps only non-optional syntax so it still runs on all Node versions.
