+++ test/configCases/plugins/define-plugin-optional-chaining/test.filter.js
@@ -1,0 +1,5 @@
+"use strict";
+
+const supportsOptionalChaining = require("../../../helpers/supportsOptionalChaining");
+
+module.exports = () => supportsOptionalChaining();
