+++ test/configCases/plugins/define-plugin-optional-chaining/webpack.config.js
@@ -1,0 +1,15 @@
+"use strict";
+
+const { DefinePlugin } = require("../../../../");
+
+/** @type {import("../../../../").Configuration} */
+module.exports = {
+	plugins: [
+		new DefinePlugin({
+			OBJECT: { SUB1: { a: 1 } },
+			"OBJECT.SUB2": { a: 1 },
+			"NOT_DEFINED.SUB2": { a: 1 },
+			STRING: JSON.stringify("string")
+		})
+	]
+};
